+++ src/TgBotLib/Exceptions/CurrencyNotSupportedException.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace TgBotLib\Exceptions;
+
+class CurrencyNotSupportedException
+{
+
+}