+++ tests/currency_test.php
@@ -1,0 +1,8 @@
+<?php
+
+
+    require 'autoload.php';
+
+    var_dump(\TgBotLib\Classes\Utilities::getCurrency('EUR'));
+    var_dump(\TgBotLib\Classes\Utilities::getCurrency('CAD'));
+    var_dump(\TgBotLib\Classes\Utilities::getCurrency('USD'));